final project/14_周守琛_吴明正_许家睿/requirements.txt
torch
numpy
pandas
open3d
pymesh
matplotlib
os
shutil
pathlib
scikit-learn
scipy
tqdm
tensorboard
easydict
pyquaternion
biopython
jupyter
torchmetrics
pdb2pqr
networkx
horovod[pytorch]
nose
